feat!: more granularity for liquidation penalty
It is now possible to set the liquidation penalty to an arbitrary number of ticks, regardless of the tick spacing.

BREAKING CHANGE: the liquidation penalty is now expressed in ticks and is stored as a uint24. This affects the `LiquidationPenaltyUpdated` event, functions `getLiquidationPenalty`, `setLiquidationPenalty`, `getLongPosition`, `getTickLiquidationPenalty`, and the `TickDat` struct

+++ src/UsdnProtocol/UsdnProtocolLong.sol
@@ -21,7 +21,7 @@ abstract contract UsdnProtocolLong is UsdnProtocolStorage, IUsdnProtocolLong {
     function getLongPosition(PositionId memory posId)
         external
         view
-        returns (Position memory pos_, uint8 liquidationPenalty_)
+        returns (Position memory pos_, uint24 liquidationPenalty_)
     {
         return Long.getLongPosition(s, posId);
     }
@@ -81,7 +81,7 @@ abstract contract UsdnProtocolLong is UsdnProtocolStorage, IUsdnProtocolLong {
     }
 
     /// @inheritdoc IUsdnProtocolLong
-    function getTickLiquidationPenalty(int24 tick) external view returns (uint8 liquidationPenalty_) {
+    function getTickLiquidationPenalty(int24 tick) external view returns (uint24 liquidationPenalty_) {
         return Long.getTickLiquidationPenalty(s, tick);
     }
 }
