+++ src/interfaces/UsdnProtocol/IUsdnProtocolFallback.sol
@@ -260,9 +260,9 @@ interface IUsdnProtocolFallback {
 
     /**
      * @notice Get the liquidation penalty applied to the liquidation price when opening a position
-     * @return The liquidation penalty (in tick spacing units)
+     * @return The liquidation penalty (in ticks)
      */
-    function getLiquidationPenalty() external view returns (uint8);
+    function getLiquidationPenalty() external view returns (uint24);
 
     /**
      * @notice Get the safety margin for the liquidation price of newly open positions
@@ -569,10 +569,10 @@ interface IUsdnProtocolFallback {
     function setMaxLeverage(uint256 newMaxLeverage) external;
 
     /**
-     * @notice Set the new liquidation penalty (in tick spacing units)
+     * @notice Set the new liquidation penalty (in ticks)
      * @param newLiquidationPenalty The new liquidation penalty
      */
-    function setLiquidationPenalty(uint8 newLiquidationPenalty) external;
+    function setLiquidationPenalty(uint24 newLiquidationPenalty) external;
 
     /**
      * @notice Set the new exponential moving average period of the funding rate
