+++ test/unit/UsdnProtocol/Actions/_PrepareClosePositionData.t.sol
@@ -115,7 +115,7 @@ contract TestUsdnProtocolActionsPrepareClosePositionData is UsdnProtocolBaseFixt
     /// @notice Assert the data in ClosePositionData depending on `isEarlyReturn`
     function _assertData(ClosePositionData memory data, bool isEarlyReturn) private view {
         uint128 currentPrice = abi.decode(currentPriceData, (uint128));
-        uint8 liquidationPenalty = protocol.getLiquidationPenalty();
+        uint24 liquidationPenalty = protocol.getLiquidationPenalty();
         uint256 positionTotalExpo = protocol.i_calcPositionTotalExpo(
             POSITION_AMOUNT,
             params.initialPrice,
